Mystery
strange errors

+++ ObjectTracking/ObjectTracking/Helper.h
@@ -6,7 +6,7 @@ using namespace cv;
 long long difTime();
 long long difTime2();
 void getRectAndCenter(const vector<Point> &contour, Point &centerC, Rect &rectC);
-int areaOf(Point &p1, Point &p2);
+double areaOf(Point &p1, Point &p2);
 Rect boundPoints(Point p1, Point p2, int exp, Point &pMinRez, Point &pMaxRez, Mat &frame);
 void minNormalRect(vector<Point> &contour, Point& min, Point &max);
 Mat cropSelectedObject(Mat & source, vector<Point> contour, bool printObjRect = false);
