+++ ObjectTracking/ObjectTracking/Kalman.cpp
@@ -27,8 +27,8 @@ void CKalmanFilter::update(Point object, Rect objectBox)
 		objV.push_back(object);
 
 		notFoundCount = 0;
-		meas.at<float>(0) = objectBox.x + objectBox.width / 2;
-		meas.at<float>(1) = objectBox.y + objectBox.height / 2;
+		meas.at<float>(0) = (float)objectBox.x + objectBox.width / 2;
+		meas.at<float>(1) = (float)objectBox.y + objectBox.height / 2;
 		meas.at<float>(2) = (float)objectBox.width;
 		meas.at<float>(3) = (float)objectBox.height;
 
@@ -63,8 +63,8 @@ void CKalmanFilter::predict(Point &objectPred, Rect &objectPredBox)
 	dT = (ticks - precTick) / cv::getTickFrequency(); //seconds
 
 	// >>>> Matrix A
-	kf.transitionMatrix.at<float>(2) = dT;
-	kf.transitionMatrix.at<float>(9) = dT;
+	kf.transitionMatrix.at<float>(2) = (float)dT;
+	kf.transitionMatrix.at<float>(9) = (float)dT;
 	// <<<< Matrix A
 
 	//std::cout << "dT: " << dT << endl;
@@ -72,13 +72,13 @@ void CKalmanFilter::predict(Point &objectPred, Rect &objectPredBox)
 	state = kf.predict();
 	//std::cout << "State post:" << endl << state << endl;
 
-	objectPredBox.width = state.at<float>(4);
-	objectPredBox.height = state.at<float>(5);
-	objectPredBox.x = state.at<float>(0) - objectPredBox.width / 2;
-	objectPredBox.y = state.at<float>(1) - objectPredBox.height / 2;
+	objectPredBox.width = (int)state.at<float>(4);
+	objectPredBox.height = (int)state.at<float>(5);
+	objectPredBox.x = (int)state.at<float>(0) - objectPredBox.width / 2;
+	objectPredBox.y = (int)state.at<float>(1) - objectPredBox.height / 2;
 
-	objectPred.x = state.at<float>(0);
-	objectPred.y = state.at<float>(1);
+	objectPred.x = (int)state.at<float>(0);
+	objectPred.y = (int)state.at<float>(1);
 
 	aproxObjV.push_back(objectPred);
 }
@@ -90,14 +90,14 @@ void CKalmanFilter::printTrajectories(Size &sz)
 	traj = Scalar(0, 0, 0);
 	if (aproxObjV.size() > 1)
 	{
-		for (int i = 0; i < aproxObjV.size() - 1; i++)
+		for (unsigned int i = 0; i < aproxObjV.size() - 1; i++)
 		{
 			line(traj, aproxObjV[i], aproxObjV[i + 1], Scalar(0, 255 * i / aproxObjV.size(), 0), 2);
 		}
 	}
 	if (objV.size() > 1)
 	{
-		for (int i = 0; i < objV.size() - 1 && objV.size(); i++)
+		for (unsigned int i = 0; i < objV.size() - 1 && objV.size(); i++)
 		{
 			line(traj, objV[i], objV[i + 1], Scalar(0, 0, 255 * i / objV.size()), 2);
 		}
@@ -132,17 +132,17 @@ CKalmanFilter::CKalmanFilter(int stateSize, int measSize, int contrSize, unsigne
 	cv::setIdentity(kf.transitionMatrix);
 
 	kf.measurementMatrix = cv::Mat::zeros(measSize, stateSize, type);
-	kf.measurementMatrix.at<float>(0) = 1.0f;
-	kf.measurementMatrix.at<float>(7) = 1.0f;
-	kf.measurementMatrix.at<float>(16) = 1.0f;
-	kf.measurementMatrix.at<float>(23) = 1.0f;
+	kf.measurementMatrix.at<float>(0) = (float)1.0f;
+	kf.measurementMatrix.at<float>(7) = (float)1.0f;
+	kf.measurementMatrix.at<float>(16) = (float)1.0f;
+	kf.measurementMatrix.at<float>(23) = (float)1.0f;
 
-	kf.processNoiseCov.at<float>(0) = 1e-2;
-	kf.processNoiseCov.at<float>(7) = 1e-2;
-	kf.processNoiseCov.at<float>(14) = 5.0f;
-	kf.processNoiseCov.at<float>(21) = 5.0f;
-	kf.processNoiseCov.at<float>(28) = 1e-2;
-	kf.processNoiseCov.at<float>(35) = 1e-2;
+	kf.processNoiseCov.at<float>(0) = (float)1e-2;
+	kf.processNoiseCov.at<float>(7) = (float)1e-2;
+	kf.processNoiseCov.at<float>(14) = (float)5.0f;
+	kf.processNoiseCov.at<float>(21) = (float)5.0f;
+	kf.processNoiseCov.at<float>(28) = (float)1e-2;
+	kf.processNoiseCov.at<float>(35) = (float)1e-2;
 
 	cv::setIdentity(kf.measurementNoiseCov, cv::Scalar(1e-1));
 }
